got structures fully working
